#3070 - Modularize webapp module
- Remove unused images
- Remove support for setting a custom logo since we no longer show a logo anyway
- Pull configuration of servlet container out of INCEpTION and into a separate configuration class
- Push most stuff from WicketApplication up into WicketApplicationBase
- Remove unused legacy WebAnno CSS stuff
