+++ inception/inception-app-webapp/src/main/java/de/tudarmstadt/ukp/inception/app/WicketApplication.properties
@@ -14,5 +14,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 product.name=INCEpTION
-warning.browser=This browser is not officially supported \u2013 please use Chrome or Safari instead.
-warning.database=INCEpTION runs with an embedded database. This is not recommended for production use \u2013 use this installation only for testing!
