+++ inception/inception-support/src/main/java/de/tudarmstadt/ukp/clarin/webanno/support/SettingsUtil.java
@@ -53,7 +53,6 @@ public class SettingsUtil
     private static final String SETTINGS_YAML_FILE = "settings.yml";
 
     public static final String CFG_LOCALE = "locale";
-    public static final String CFG_STYLE_LOGO = "style.logo";
     public static final String CFG_AUTH_MODE = "auth.mode";
     public static final String CFG_AUTH_PREAUTH_NEWUSER_ROLES = "auth.preauth.newuser.roles";
     public static final String CFG_WARNINGS_EMBEDDED_DATABASE = "warnings.embeddedDatabase";
